feat: Added login authentication + user management Note that after the client creates an account, they will still need to log in, because the JWT token is only generated after a successful login.
JWT authentication has been implemented for authentication and authorisation of users.
JWT tokens are being passed around using HTTP-only cookies, which are set on the client side after a successful login.
jwt.ts contains the logic for verifying the JWT token and extracting and setting the user_id from it.

Transactions + Recurring Transactions all uses user_id to ensure that the items are associated with the correct user.

index.d.ts has been updated to include the user_id in the request object, so that it can be accessed in the controllers.